+++ .gitignore
@@ -45,7 +45,8 @@ coverage.xml
 
 # Django stuff:
 *.log
-local_settings.py
+settings/local.py
+settings/local_code.py
 
 # Sphinx documentation
 docs/_build/
